Restructured files and more (description)
Restructured the position of files and added a Graph directory inside the API directory to differentiate in API's in the future.

Implemented more possible graph calls that people can use when using the Graph.cs class object.

Added more test graph call code to the Test Token button so people can see an example on how to call the api.

Added "Files.Read.All" as a new permission so the api has access to read files that are inside a OneDrive.