New control.
Added the Reload video control. It is designed to be used when you cannot seek the video anymore duo to a OneDrive or LibVLC problem.

This is an attempt to make the issue as less annoying as possible. It reloards the video and starts back at the time you where before the reload.