Removed MVVM Light
Removed the MVVMLight package as it is deprecated and added the
Microsoft.Toolkit.Mvvm package as the successor. Some implementations
had to be redone as MVVM Toolkit didn't provide a drop in replacement.
An example for such case is the clearing/ unregistering of a viewmodel.
I have added a method in the code behind of the main page that checks if
it will navigate to the login page. If that is the case, it clears the
navigation cache resulting in all cached pages being cleared (except for
the pages using the 'required' navigation cache mode). The reason I
added the cache mode enabled to the main page is so that users don't
have to navigato through their directories after navigation back to the
main page after every navigation.