fix: Update volume when creating LibVLC object
With MVVMLight, the static volumeUpdated was initialized to false on
every navigation action to the page. With the new navigation, the
static volumeUpdated actually held it's value which resulted in the
volume not being updated after navigating to the page for a second
time. This has been fixed by making the variable not static, so it
is initialized every time correctly.

Also added an assertion to self-document this, and added clarification
comment.